Create of AirplaneServiceTest.java, PilotServiceTest.java, FlightServiceTest.java

+++ business/src/main/java/br/com/embraer/flights/business/dto/FlightDto.java
@@ -14,8 +14,8 @@ public class FlightDto {
 	private String status;
 	private String pilotName;
 	private String airplaneName;
-	private Integer pilotId;
-	private Integer airplaneId;
+	private Long pilotId;
+	private Long airplaneId;
 
 
 }
